+++ TaskExplorer/GUI/TaskExplorer.h
@@ -352,8 +352,10 @@ private:
 	CCustomItemDelegate* m_pCustomItemDelegate;
 
 	void				LoadLanguage();
-	QTranslator			m_Translator;
-	QByteArray			m_Translation;
+	void				LoadLanguage(const QString& Lang, const QString& Module, int Index);
+	QTranslator			m_Translator[2];
+	QString				m_Language;
+	//quint32				m_LanguageId;
 };
 
 extern CTaskExplorer*	theGUI;